Make Shrine::DataFile truly top-level citizen
